Laboratorium10/pyproject.toml
[tool.poetry]
name = "lab10"
version = "0.1.0"
description = ""
authors = ["Jakub Wlodarczyk"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.8,<=3.10"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
Pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
